Split layout tests into separate modules.
